backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data validation and serialization
pydantic==2.5.0

# Development and testing
pytest==7.4.3
httpx==0.25.2

# System Monitoring
psutil==5.9.6

# AI Processing Pipeline
openai>=1.0.0
pytesseract>=0.3.10
Pillow>=9.0.0
SpeechRecognition>=3.10.0
pydub>=0.25.1
docx2txt>=0.8
PyPDF2>=3.0.0

# Background Processing
APScheduler>=3.10.0
celery>=5.3.0
redis>=4.5.0

# Additional Audio Processing
ffmpeg-python>=0.2.0
librosa>=0.10.0

# Image Processing Enhancements
opencv-python>=4.7.0
numpy>=1.21.0

# Document Processing Enhancements
python-docx>=0.8.11
PyMuPDF>=1.22.0

# Environment and Configuration
python-dotenv>=1.0.0

# Excel / Analytics
pandas>=1.5.0
openpyxl>=3.0.10
xlsxwriter>=3.0.3

extract-msg==0.29.*
